Adds shutdown command and various fixes
Adds a shutdown command
Adds a force-config-reset command
Adds a help command
Adds a test command
Bot is now assumed to talk in the server instead of DMs though DMs sort of still work.

When a force-config is not provided fall back to the existing request preferences.

force-config spits out what it changed.